endret type på beregningsgrunnlag

+++ app/main/api/Maksimum.kt
@@ -19,7 +19,7 @@ data class Vedtak(
     val vedtaksdato: String, //reg_dato
     val periode: Periode,
     val rettighetsType: String, ////aktivitetsfase //Aktfasekode
-    val beregningsgrunnlag: Float
+    val beregningsgrunnlag: Int
 )
 
 data class UtbetalingMedMer(
